requirements.txt
# Core dependencies
streamlit==1.39.0
yfinance==0.2.48
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
plotly==5.24.1
scipy==1.14.1
openai==1.54.5
langgraph==0.2.49
langchain-core==0.3.19
python-dotenv==1.0.1

# Enhancements - New dependencies for extended functionality
duckdb==0.9.2           # Database for caching and pattern statistics
pandas_ta==0.3.14b0     # Enhanced technical indicators
requests==2.31.0        # For AlphaVantage API
pytest==7.4.3           # Testing framework
black==23.12.1          # Code formatter
ruff==0.1.9            # Linter
pytest-cov==4.1.0      # Code coverage for pytest